chore: move persistence pkg back to controller.
